[SEARCH-1430] abstract alfresco class refactor
