+++ alfresco-search/src/test/java/org/alfresco/solr/AbstractAlfrescoDistributedTest.java
@@ -5,30 +5,24 @@ import org.alfresco.solr.client.Node;
 import org.alfresco.solr.client.NodeMetaData;
 import org.alfresco.solr.client.SOLRAPIQueueClient;
 import org.alfresco.solr.client.Transaction;
-import org.alfresco.solr.util.SolrResponsesComparator;
-import org.apache.commons.io.FileUtils;
+import org.alfresco.solr.basics.RandomSupplier;
+import org.alfresco.solr.basics.SolrResponsesComparator;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.index.Term;
-import org.apache.lucene.search.*;
-import org.apache.solr.SolrTestCaseJ4;
+import org.apache.lucene.search.Query;
+import org.apache.lucene.search.TermQuery;
 import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.SolrRequest;
-import org.apache.solr.client.solrj.SolrResponse;
-import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.embedded.JettyConfig;
+import org.apache.solr.client.solrj.SolrRequest;
+import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.embedded.JettySolrRunner;
-import org.apache.solr.client.solrj.embedded.SSLConfig;
-import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.apache.solr.client.solrj.io.Tuple;
 import org.apache.solr.client.solrj.io.stream.TupleStream;
 import org.apache.solr.client.solrj.request.QueryRequest;
 import org.apache.solr.client.solrj.request.UpdateRequest;
 import org.apache.solr.client.solrj.response.QueryResponse;
 import org.apache.solr.client.solrj.response.UpdateResponse;
-import org.apache.solr.common.SolrDocument;
-import org.apache.solr.common.SolrDocumentList;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.params.CoreAdminParams;
 import org.apache.solr.common.params.ModifiableSolrParams;
@@ -39,26 +33,37 @@ import org.apache.solr.core.SolrCore;
 import org.apache.solr.request.LocalSolrQueryRequest;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
-import org.apache.solr.update.AddUpdateCommand;
-import org.eclipse.jetty.servlet.ServletHolder;
-import org.junit.Assert;
+import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.rules.ExternalResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.*;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.*;
+import java.io.IOException;
 import java.lang.invoke.MethodHandles;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Base64;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
+import java.util.Random;
 
-import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.*;
-import static org.alfresco.solr.AlfrescoSolrUtils.createCoreUsingTemplate;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_ACLID;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_DBID;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_DOC_TYPE;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_INTXID;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_OWNER;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_SOLR4_ID;
+import static org.alfresco.repo.search.adaptor.lucene.QueryConstants.FIELD_VERSION;
 
 /**
  * Clone of a helper base class for distributed search test cases
@@ -82,23 +87,22 @@ import static org.alfresco.solr.AlfrescoSolrUtils.createCoreUsingTemplate;
  * @author Michael Suzuki 
  */
 @ThreadLeakLingering(linger = 5000)
-public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
+public abstract class AbstractAlfrescoDistributedTest extends SolrTestEnvironmentInitializer
 {
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     
-    // TODO: this shouldn't be static. get the random when you need it to avoid
-    // sharing.
-    public static Random r;
-    private AtomicInteger nodeCnt = new AtomicInteger(0);
-    protected boolean useExplicitNodeNames;
+    protected String[] deadServers;
+    protected SolrResponsesComparator solrComparator = new SolrResponsesComparator();
+    protected RandomSupplier solrRandomSupplier;
 
-    public static Properties DEFAULT_CORE_PROPS = new Properties();
-
-    static {
-        DEFAULT_CORE_PROPS.setProperty("alfresco.commitInterval", "1000");
-        DEFAULT_CORE_PROPS.setProperty("alfresco.newSearcherInterval", "2000");
-    }
-
+    // to stress with higher thread counts and requests, make sure the junit
+    // xml formatter is not being used (all output will be buffered before
+    // transformation to xml and cause an OOM exception).
+    protected int stress = TEST_NIGHTLY ? 2 : 0;
+    protected boolean verifyStress = true;
+    protected int nThreads = 3;
+    protected String id = "id";
+    
     /**
      * Set's the value of the "hostContext" system property to a random path
      * like string (which may or may not contain sub-paths). This is used in the
@@ -111,123 +115,27 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
      *
      */
     @BeforeClass
-    public static void setup()
+    public static void setUpSolrTestProperties()
     {
         System.setProperty("alfresco.test", "true");
         System.setProperty("solr.tests.maxIndexingThreads", "10");
         System.setProperty("solr.tests.ramBufferSizeMB", "1024");
-        testDir = new File(System.getProperty("user.dir") + "/target/jettys");
-        r = new Random(random().nextLong());
-
     }
 
-    protected Map<String, JettySolrRunner> jettyContainers = new HashMap<>();
-    protected Map<String, SolrClient> collectionToStandaloneClient = new HashMap<>();
-
-    protected List<JettySolrRunner> jettyShards = new ArrayList<>();
-    protected List<SolrClient> clientShards = new ArrayList<>();
-
-    protected String[] deadServers;
-    protected String shards;
-    protected String[] shardsArr;
-
-    protected static File testDir;
-    protected SolrResponsesComparator solrComparator = new SolrResponsesComparator();
-
-    // to stress with higher thread counts and requests, make sure the junit
-    // xml formatter is not being used (all output will be buffered before
-    // transformation to xml and cause an OOM exception).
-    protected int stress = TEST_NIGHTLY ? 2 : 0;
-    protected boolean verifyStress = true;
-    protected int nThreads = 3;
-
-    protected int clientConnectionTimeout = DEFAULT_CONNECTION_TIMEOUT;
-    protected int clientSoTimeout = 90000;
-
-    protected String id = "id";
-
-    public static RandVal rint = new RandVal()
+    @Before
+    public void setupPerTest()
     {
-        @Override
-        public Object val()
-        {
-            return r.nextInt();
-        }
-    };
+        this.solrRandomSupplier = new RandomSupplier();
+    }
+    
 
-    public static RandVal rlong = new RandVal()
-    {
-        @Override
-        public Object val()
-        {
-            return r.nextLong();
-        }
-    };
-
-    public static RandVal rfloat = new RandVal()
-    {
-        @Override
-        public Object val()
-        {
-            return r.nextFloat();
-        }
-    };
-
-    public static RandVal rdouble = new RandVal()
-    {
-        @Override
-        public Object val()
-        {
-            return r.nextDouble();
-        }
-    };
-
-    public static RandVal rdate = new RandDate();
-
-    public static String[] fieldNames = new String[]
+    public String[] fieldNames = new String[]
     { "n_ti1", "n_f1", "n_tf1", "n_d1", "n_td1", "n_l1", "n_tl1", "n_dt1", "n_tdt1" };
-    public static RandVal[] randVals = new RandVal[]
-    { rint, rfloat, rfloat, rdouble, rdouble, rlong, rlong, rdate, rdate };
+    
 
     protected String[] getFieldNames()
     {
         return fieldNames;
-    }
-
-    protected RandVal[] getRandValues()
-    {
-        return randVals;
-    }
-
-    /**
-     * Subclasses can override this to change a test's solr home (default is in
-     * test-files)
-     */
-    public String getTestFilesHome()
-    {
-        return System.getProperty("user.dir") + "/target/test-classes/test-files";
-    }
-    public void distribSetUp(String serverName) throws Exception
-    {
-        SolrTestCaseJ4.resetExceptionIgnores(); // ignore anything with
-                                                // ignore_exception in it
-        System.setProperty("solr.test.sys.prop1", "propone");
-        System.setProperty("solr.test.sys.prop2", "proptwo");
-        System.setProperty("solr.directoryFactory", "org.apache.solr.core.MockDirectoryFactory");
-        System.setProperty("solr.log.dir", testDir.toPath().resolve(serverName).toString());
-    }
-
-    public void distribTearDown() throws Exception
-    {
-        System.clearProperty("solr.directoryFactory");
-        System.clearProperty("solr.log.dir");
-
-        SOLRAPIQueueClient.nodeMetaDataMap.clear();
-        SOLRAPIQueueClient.transactionQueue.clear();
-        SOLRAPIQueueClient.aclChangeSetQueue.clear();
-        SOLRAPIQueueClient.aclReadersMap.clear();
-        SOLRAPIQueueClient.aclMap.clear();
-        SOLRAPIQueueClient.nodeMap.clear();
     }
 
     protected void putHandleDefaults() {
@@ -244,9 +152,6 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
      */
     public void waitForDocCountAllCores(Query query, int count, long waitMillis) throws Exception
     {
-        List<SolrCore> cores = getJettyCores(jettyContainers.values());
-        cores.addAll(getJettyCores(jettyShards));
-
         long begin = System.currentTimeMillis();
 
         for (SolrClient client : getStandaloneAndShardedClients()) {
@@ -256,14 +161,12 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
 
     public void waitForDocCountAllShards(Query query, int count, long waitMillis) throws Exception
     {
-        List<SolrCore> cores = getJettyCores(jettyShards);
         long begin = System.currentTimeMillis();
         for (SolrClient client : getShardedClients()) {
             waitForDocCountCore(client, luceneToSolrQuery(query), count, waitMillis, begin);
         }
     }
-
-
+    
     /**
      * Delele by query on all Clients
      *
@@ -286,8 +189,9 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
      */
     protected SolrClient getDefaultTestClient()
     {
-        return collectionToStandaloneClient.get(DEFAULT_TEST_CORENAME);
+        return solrCollectionNameToStandaloneClient.get(DEFAULT_TEST_CORENAME);
     }
+    
 
     protected List<SolrClient> getShardedClients()
     {
@@ -302,7 +206,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     public List<SolrClient> getStandaloneAndShardedClients()
     {
         List<SolrClient> clients = new ArrayList();
-        clients.addAll(collectionToStandaloneClient.values());
+        clients.addAll(solrCollectionNameToStandaloneClient.values());
         clients.addAll(clientShards);
         return clients;
     }
@@ -311,7 +215,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     public List<SolrClient> getStandaloneClients()
     {
         List<SolrClient> clients = new ArrayList();
-        clients.addAll(collectionToStandaloneClient.values());
+        clients.addAll(solrCollectionNameToStandaloneClient.values());
         return clients;
     }
 
@@ -410,8 +314,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     }
 
     protected void injectDocToShards(long txnId, long aclId, long dbId, String owner) throws Exception {
-        List<SolrCore> cores = getJettyCores(jettyShards);
-        for(SolrCore core : cores) {
+        for(SolrClient clientShard : clientShards) {
             SolrInputDocument doc = new SolrInputDocument();
             String id = AlfrescoSolrDataModel.getNodeDocumentId(AlfrescoSolrDataModel.DEFAULT_TENANT, aclId, dbId);
             doc.addField(FIELD_SOLR4_ID, id);
@@ -420,21 +323,8 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
             doc.addField(FIELD_INTXID, "" + txnId);
             doc.addField(FIELD_ACLID, "" + aclId);
             doc.addField(FIELD_OWNER, owner);
-
-            AbstractAlfrescoSolrTests.SolrServletRequest solrQueryRequest = null;
-            try
-            {
-                solrQueryRequest = new AbstractAlfrescoSolrTests.SolrServletRequest(core, null);
-
-                AddUpdateCommand addDocCmd = new AddUpdateCommand(solrQueryRequest);
-                addDocCmd.overwrite = true;
-                addDocCmd.solrDoc = doc;
-                core.getUpdateHandler().addDoc(addDocCmd);
-            }
-            finally
-            {
-                solrQueryRequest.close();
-            }
+            clientShard.add(doc);
+            clientShard.commit();
         }
     }
 
@@ -473,13 +363,10 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     {
         int shardHit = 0;
         List<SolrClient> clients = getShardedClients();
-        List<SolrCore> cores = getJettyCores(jettyShards);
         SolrQuery query = luceneToSolrQuery(new TermQuery(new Term(FIELD_DOC_TYPE, SolrInformationServer.DOC_TYPE_NODE)));
         StringBuilder error = new StringBuilder();
         for (int i = 0; i < clients.size(); ++i)
         {
-
-            SolrCore core = cores.get(i);
             SolrClient client = clients.get(i);
 
 
@@ -495,15 +382,15 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
             {
                 if (ignoreZero && totalHits == 0)
                 {
-                    log.info(core.getName()+": have zero hits ");
+                    log.info(client+": have zero hits ");
                 }
                 else
                 {
-                    error.append(" "+core.getName()+": ");
+                    error.append(" "+client+": ");
                     error.append("Expected nodes per shard greater than "+count+" found "+totalHits+" : "+query.toString());
                 }
             }
-            log.info(core.getName()+": Hits "+totalHits);
+            log.info(client+": Hits "+totalHits);
 
         }
 
@@ -586,156 +473,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         }
         throw new Exception("Core:Wait error expected "+expectedNumFound+" found "+totalHits+" : "+query.toString());
     }
-
-    /**
-     * Creates a JettySolrRunner (if one didn't exist already). DOES NOT START IT.
-     * @return
-     * @throws Exception
-     */
-    private JettySolrRunner createJetty(String jettyKey, boolean basicAuth) throws Exception
-    {
-        if (jettyContainers.containsKey(jettyKey))
-        {
-            return jettyContainers.get(jettyKey);
-        }
-        else
-        {
-            Path jettySolrHome = testDir.toPath().resolve(jettyKey);
-            seedSolrHome(jettySolrHome);
-            JettySolrRunner jetty = createJetty(jettySolrHome.toFile(), null, null, false, getSchemaFile(), basicAuth);
-            return jetty;
-        }
-    }
-
-    /**
-     * Adds the core config information to the jetty file system.
-     * Its best to call this before calling start() on Jetty
-     * @param jettyKey
-     * @param sourceConfigName
-     * @param coreName
-     * @param additionalProperties
-     * @throws Exception
-     */
-    private void addCoreToJetty(String jettyKey, String sourceConfigName, String coreName, Properties additionalProperties) throws Exception
-    {
-        Path jettySolrHome = testDir.toPath().resolve(jettyKey);
-        Path coreSourceConfig = new File(getTestFilesHome() + "/"+sourceConfigName).toPath();
-        Path coreHome = jettySolrHome.resolve(coreName);
-        seedCoreDir(coreName, coreSourceConfig, coreHome);
-        updateSolrCoreProperties(coreHome, additionalProperties);
-    }
-
-
-    private void updateSolrCoreProperties(Path coreHome, Properties additionalProperties) throws IOException
-    {
-        if(additionalProperties != null)
-        {
-            InputStream in = null;
-            OutputStream out = null;
-            try
-            {
-                Properties properties = new Properties();
-                String solrcoreProperties = coreHome.resolve("conf/solrcore.properties").toString();
-                in = new FileInputStream(solrcoreProperties);
-                properties.load(in);
-                in.close();
-                additionalProperties.entrySet().forEach(x-> properties.put(x.getKey(),x.getValue()));
-                out = new FileOutputStream(solrcoreProperties);
-                properties.store(out, null);
-            }
-            finally
-            {
-                out.close();
-                in.close();
-            }
-            
-        }
-        
-    }
-
-    /**
-     * Starts jetty if its not already running
-     * @param jsr
-     * @throws Exception
-     */
-    private void startJetty(JettySolrRunner jsr) throws Exception {
-        if (!jsr.isRunning())
-        {
-            jsr.start();
-        }
-    }
-
-    protected void createServers(String jettyKey, String[] coreNames, int numShards, Properties additionalProperties) throws Exception
-    {
-        boolean basicAuth = additionalProperties != null ? Boolean.parseBoolean(additionalProperties.getProperty("BasicAuth", "false")) : false;
-
-        JettySolrRunner jsr =  createJetty(jettyKey, basicAuth);
-        jettyContainers.put(jettyKey, jsr);
-
-        Properties properties = new Properties();
-
-        if(additionalProperties != null && additionalProperties.size() > 0) {
-            properties.putAll(additionalProperties);
-            properties.remove("shard.method");
-        }
-
-        for (int i = 0; i < coreNames.length; i++) 
-        {
-            addCoreToJetty(jettyKey, coreNames[i], coreNames[i], properties);
-        }
-
-        //Now start jetty
-        startJetty(jsr);
-
-        int jettyPort = jsr.getLocalPort();
-        for (int i = 0; i < coreNames.length; i++)
-        {
-            String url = buildUrl(jettyPort) + "/" + coreNames[i];
-            log.info(url);
-            collectionToStandaloneClient.put(coreNames[i], createNewSolrClient(url));
-        }
-
-        shardsArr = new String[numShards];
-        StringBuilder sb = new StringBuilder();
-
-        if (additionalProperties == null) {
-            additionalProperties = new Properties();
-        }
-
-
-        String[] ranges = {"0-100", "100-200", "200-300", "300-400"};
-
-
-
-        for (int i = 0; i < numShards; i++)
-        {
-            Properties props = new Properties();
-            props.putAll(additionalProperties);
-            if (sb.length() > 0) sb.append(',');
-            final String shardname = "shard" + i;
-            props.put("shard.instance", Integer.toString(i));
-            props.put("shard.count", Integer.toString(numShards));
-
-            if("DB_ID_RANGE".equalsIgnoreCase(props.getProperty("shard.method")))
-            {
-                props.put("shard.range", ranges[i]);
-            }
-
-            String shardKey = jettyKey+"_shard_"+i;
-            JettySolrRunner j =  createJetty(shardKey, basicAuth);
-            //use the first corename specified as the Share template
-            addCoreToJetty(shardKey, coreNames[0], shardname, props);
-            jettyShards.add(j);
-            startJetty(j);
-            String shardStr = buildUrl(j.getLocalPort()) + "/" + shardname;
-            log.info(shardStr);
-            SolrClient clientShard = createNewSolrClient(shardStr);
-            clientShards.add(clientShard);
-            shardsArr[i] = shardStr;
-            sb.append(shardStr);
-        }
-        shards = sb.toString();
-    }
+    
 
     protected void setDistributedParams(ModifiableSolrParams params)
     {
@@ -767,23 +505,9 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         return tuples;
     }
 
-    protected String getShardsString(List<SolrClient> clientList)
-    {
-        StringBuilder buf = new StringBuilder();
-        for(int i=0; i<clientList.size(); ++i) {
-            HttpSolrClient solrClient = (HttpSolrClient)clientList.get(i);
-
-            if(buf.length() > 0) {
-                buf.append(",");
-            }
-
-            buf.append(solrClient.getBaseURL());
-        }
-        return buf.toString();
-    }
-
     protected String getShardsString()
     {
+        Random r = solrRandomSupplier.getRandomGenerator();
         if (deadServers == null)
             return shards;
 
@@ -812,135 +536,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
 
         return sb.toString();
     }
-
-    protected void destroyServers() throws Exception
-    {
-        List<String> solrHomes = new ArrayList<String>();
-        for (JettySolrRunner jetty : jettyContainers.values())
-        {
-            solrHomes.add(jetty.getSolrHome());
-            jetty.stop();
-        }
-        for (SolrClient jClients : collectionToStandaloneClient.values())
-        {
-            jClients.close();
-        }
-
-        for (JettySolrRunner jetty : jettyShards)
-        {
-            solrHomes.add(jetty.getSolrHome());
-            jetty.stop();
-        }
-
-        for (SolrClient client : clientShards)
-        {
-            client.close();
-        }
-
-        for(String home : solrHomes)
-        {
-            FileUtils.deleteDirectory(new File(home, "ContentStore"));
-        }
-
-        clientShards.clear();
-        jettyShards.clear();
-        jettyContainers.clear();
-        collectionToStandaloneClient.clear();
-    }
-
-    public JettySolrRunner createJetty(File solrHome, String dataDir, String shardList, boolean sslEnabled, String schemaOverride, boolean basicAuth) throws Exception
-    {
-        return createJetty(solrHome, dataDir, shardList, sslEnabled, schemaOverride, useExplicitNodeNames, basicAuth);
-    }
-
-    /**
-     * Create a solr jetty server.
-     * 
-     * @param solrHome
-     * @param dataDir
-     * @param shardList
-     * @param schemaOverride
-     * @param explicitCoreNodeName
-     * @return
-     * @throws Exception
-     */
-    public JettySolrRunner createJetty(File solrHome, String dataDir, String shardList, boolean sslEnabled,
-            String schemaOverride, boolean explicitCoreNodeName, boolean basicAuth) throws Exception
-    {
-        Properties props = new Properties();
-        if (schemaOverride != null)
-            props.setProperty("schema", schemaOverride);
-        if (shardList != null)
-            props.setProperty("shards", shardList);
-        if (dataDir != null)
-        {
-            props.setProperty("solr.data.dir", dataDir);
-        }
-
-        if (explicitCoreNodeName)
-        {
-            props.setProperty("coreNodeName", Integer.toString(nodeCnt.incrementAndGet()));
-        }
-        SSLConfig sslConfig = new SSLConfig(sslEnabled, false, null, null, null, null);
-
-        JettyConfig config = null;
-
-        if(basicAuth) {
-            System.out.println("###### adding basic auth ######");
-            config = JettyConfig.builder().setContext("/solr").withFilter(BasicAuthFilter.class, "/sql/*").stopAtShutdown(true).withSSLConfig(sslConfig).build();
-        } else {
-            System.out.println("###### no basic auth ######");
-            config = JettyConfig.builder().setContext("/solr").stopAtShutdown(true).withSSLConfig(sslConfig).build();
-        }
-
-        JettySolrRunner jetty = new JettySolrRunner(solrHome.getAbsolutePath(), props, config);
-        // .stopAtShutdown(true)
-        // .withFilters(getExtraRequestFilters())
-        // .withServlets(getExtraServlets())
-        // .withSSLConfig(sslConfig)
-        // .build());
-        return jetty;
-    }
-
-    /**
-     * Override this method to insert extra servlets into the JettySolrRunners
-     * that are created using createJetty()
-     */
-    public SortedMap<ServletHolder, String> getExtraServlets()
-    {
-        return null;
-    }
-
-    /**
-     * Override this method to insert extra filters into the JettySolrRunners
-     * that are created using createJetty()
-     */
-    public SortedMap<Class<? extends Filter>, String> getExtraRequestFilters()
-    {
-        return null;
-    }
-
-    protected SolrClient createNewSolrClient(String url)
-    {
-        try
-        {
-            // setup the client...
-            HttpSolrClient client = new HttpSolrClient(url);
-            client.setConnectionTimeout(clientConnectionTimeout);
-            client.setSoTimeout(clientSoTimeout);
-            client.setDefaultMaxConnectionsPerHost(100);
-            client.setMaxTotalConnections(100);
-            return client;
-        } catch (Exception ex)
-        {
-            throw new RuntimeException(ex);
-        }
-    }
-
-    protected String buildUrl(int port)
-    {
-        return buildUrl(port, "/solr");
-    }
+    
 
     protected void addFields(SolrInputDocument doc, Object... fields)
     {
@@ -961,7 +557,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
 
     protected SolrInputDocument addRandFields(SolrInputDocument sdoc)
     {
-        addFields(sdoc, getRandFields(getFieldNames(), getRandValues()));
+        addFields(sdoc, solrRandomSupplier.getRandFields(getFieldNames(), solrRandomSupplier.getRandValues()));
         return sdoc;
     }
 
@@ -1092,9 +688,9 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
 
     }
 
-    protected QueryResponse queryServer(ModifiableSolrParams params) throws SolrServerException, IOException
+    protected QueryResponse queryRandomShard(ModifiableSolrParams params) throws SolrServerException, IOException
     {
-        // query a random server
+        Random r = solrRandomSupplier.getRandomGenerator();
         int which = r.nextInt(clientShards.size());
         SolrClient client = clientShards.get(which);
         QueryResponse rsp = client.query(params);
@@ -1111,7 +707,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         {
             params.remove("distrib");
             setDistributedParams(params);
-            QueryResponse rsp = queryServer(json, params);
+            QueryResponse rsp = queryRandomShard(json, params);
             System.out.println("Cluster Response:"+rsp);
             System.out.println("Control Response:"+controlRsp);
             solrComparator.compareResponses(rsp, controlRsp);
@@ -1123,9 +719,9 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         }
     }
 
-    protected QueryResponse queryServer(String json, SolrParams params) throws SolrServerException, IOException
+    protected QueryResponse queryRandomShard(String json, SolrParams params) throws SolrServerException, IOException
     {
-        // query a random server
+        Random r = solrRandomSupplier.getRandomGenerator();
         int which = r.nextInt(clientShards.size());
         SolrClient client = clientShards.get(which);
         QueryRequest request = getAlfrescoRequest(json, params);
@@ -1139,11 +735,11 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     }
 
     /**
-     * Returns the QueryResponse from {@link #queryServer}
+     * Returns the QueryResponse from {@link #queryRandomShard}
      */
     protected QueryResponse query(SolrClient solrClient, boolean setDistribParams, SolrParams p) throws Exception
     {
-
+        Random r = solrRandomSupplier.getRandomGenerator();
         final ModifiableSolrParams params = new ModifiableSolrParams(p);
 
         // TODO: look into why passing true causes fails
@@ -1155,7 +751,7 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         if (setDistribParams)
             setDistributedParams(params);
 
-        QueryResponse rsp = queryServer(params);
+        QueryResponse rsp = queryRandomShard(params);
 
         solrComparator.compareResponses(rsp, controlRsp);
 
@@ -1224,112 +820,6 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
         return first;
     }
 
-    public static Object[] getRandFields(String[] fields, RandVal[] randVals)
-    {
-        Object[] o = new Object[fields.length * 2];
-        for (int i = 0; i < fields.length; i++)
-        {
-            o[i * 2] = fields[i];
-            o[i * 2 + 1] = randVals[i].uval();
-        }
-        return o;
-    }
-
-    public static abstract class RandVal
-    {
-        public static Set uniqueValues = new HashSet();
-
-        public abstract Object val();
-
-        public Object uval()
-        {
-            for (;;)
-            {
-                Object v = val();
-                if (uniqueValues.add(v))
-                    return v;
-            }
-        }
-    }
-
-    public static class RandDate extends RandVal
-    {
-        @Override
-        public Object val()
-        {
-            long v = r.nextLong();
-            Date d = new Date(v);
-            return d.toInstant().toString();
-        }
-    }
-
-    protected String getSolrXml()
-    {
-        return "solr.xml";
-    }
-
-    /**
-     * Given a directory that will be used as the SOLR_HOME for a jetty
-     * instance, seeds that directory with the contents of {@link #getTestFilesHome}
-     * and ensures that the proper {@link #getSolrXml} file is in place.
-     */
-    protected void seedSolrHome(Path jettyHome) throws IOException
-    {
-        String solrxml = getSolrXml();
-        if (solrxml != null)
-        {
-            FileUtils.copyFile(new File(getTestFilesHome(), solrxml), jettyHome.resolve(getSolrXml()).toFile());
-        }
-        //Add solr home conf folder with alfresco based configuration.
-        FileUtils.copyDirectory(new File(getTestFilesHome() + "/conf"), jettyHome.resolve("conf").toFile());
-        // Add alfresco data model def
-        FileUtils.copyDirectory(new File(getTestFilesHome() + "/alfrescoModels"), jettyHome.resolve("alfrescoModels").toFile());
-        // Add templates
-        FileUtils.copyDirectory(new File(getTestFilesHome() + "/templates"), jettyHome.resolve("templates").toFile());
-        //add solr alfresco properties
-        FileUtils.copyFile(new File(getTestFilesHome() + "/log4j-solr.properties"), jettyHome.resolve("log4j-solr.properties").toFile());
-        
-    }
-
-    /**
-     * Given a directory that will be used as the <code>coreRootDirectory</code>
-     * for a jetty instance, Creates a core directory named
-     * {@link #DEFAULT_TEST_CORENAME} using a trivial
-     * <code>core.properties</code> if this file does not already exist.
-     *
-     * @see #writeCoreProperties(Path,String)
-     * @see #CORE_PROPERTIES_FILENAME
-     */
-    private void seedCoreDir(String coreName, Path coreSourceConfig, Path coreDirectory) throws IOException
-    {
-        //Prepare alfresco solr core.
-        Path confDir = coreDirectory.resolve("conf");
-        confDir.toFile().mkdirs();
-        if (Files.notExists(coreDirectory.resolve(CORE_PROPERTIES_FILENAME)))
-        {
-            Properties coreProperties = new Properties();
-            coreProperties.setProperty("name", coreName);
-            writeCoreProperties(coreDirectory, coreProperties, this.getTestName());
-        } // else nothing to do, DEFAULT_TEST_CORENAME already exists
-        //Add alfresco solr configurations
-        FileUtils.copyDirectory(coreSourceConfig.resolve("conf").toFile(), confDir.toFile());
-    }
-
-    protected void setupJettySolrHome(String coreName, Path jettyHome) throws IOException
-    {
-        seedSolrHome(jettyHome);
-
-        Properties coreProperties = new Properties();
-        coreProperties.setProperty("name", coreName);
-        coreProperties.setProperty("shard", "${shard:}");
-        coreProperties.setProperty("collection", "${collection:" + coreName + "}");
-        coreProperties.setProperty("config", "${solrconfig:solrconfig.xml}");
-        coreProperties.setProperty("schema", "${schema:schema.xml}");
-        coreProperties.setProperty("coreNodeName", "${coreNodeName:}");
-
-        writeCoreProperties(jettyHome.resolve("cores").resolve(coreName), coreProperties, coreName);
-    }
-
     public void indexTransaction(Transaction transaction, List<Node> nodes, List<NodeMetaData> nodeMetaDatas)
     {
         //First map the nodes to a transaction.
@@ -1392,8 +882,8 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
 
     public SolrQueryResponse rangeCheck(int shard) throws Exception {
         while(true) {
-            List<SolrCore> cores = getJettyCores(jettyShards);
-            List<AlfrescoCoreAdminHandler> alfrescoCoreAdminHandlers = getAdminHandlers(jettyShards);
+            List<SolrCore> cores = getJettyCores(solrShards);
+            List<AlfrescoCoreAdminHandler> alfrescoCoreAdminHandlers = getAdminHandlers(solrShards);
             SolrCore core = cores.get(shard);
             AlfrescoCoreAdminHandler alfrescoCoreAdminHandler = alfrescoCoreAdminHandlers.get(shard);
             SolrQueryResponse response = callHandler(alfrescoCoreAdminHandler, core, "RANGECHECK");
@@ -1408,8 +898,8 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     }
 
     public SolrQueryResponse expand(int shard, int value) {
-        List<SolrCore> cores = getJettyCores(jettyShards);
-        List<AlfrescoCoreAdminHandler> alfrescoCoreAdminHandlers = getAdminHandlers(jettyShards);
+        List<SolrCore> cores = getJettyCores(solrShards);
+        List<AlfrescoCoreAdminHandler> alfrescoCoreAdminHandlers = getAdminHandlers(solrShards);
         SolrCore core = cores.get(shard);
         AlfrescoCoreAdminHandler alfrescoCoreAdminHandler = alfrescoCoreAdminHandlers.get(shard);
         SolrQueryResponse response = callExpand(alfrescoCoreAdminHandler, core, value);
@@ -1431,147 +921,6 @@ public abstract class AbstractAlfrescoDistributedTest extends SolrTestCaseJ4
     private long getFieldValueLong(Document doc, String fieldName)
     {
         return Long.parseLong(getFieldValueString(doc, fieldName));
-    }
-
-    /**
-     * A JUnit Rule to setup Jetty
-     */
-    public class JettyServerRule extends ExternalResource
-    {
-
-        final String serverName;
-        final String[] coreNames;
-        final int numShards;
-        final Properties solrcoreProperties;
-
-        /**
-         * Creates the jetty servers
-         * @param serverName The key to use for the Jetty server name
-         * @param numShards Number of shards required
-         * @param solrcoreProperties Additional properties to add to the solrcore.properties
-         * @param coreNames names of the core config folders
-         */
-        public JettyServerRule(String serverName, int numShards, Properties solrcoreProperties, String ...coreNames)
-        {
-            this.serverName = serverName;
-            this.coreNames = coreNames == null?new String[0]:coreNames;
-            this.numShards = numShards;
-            this.solrcoreProperties = solrcoreProperties;
-        }
-
-        /**
-         * Creates the jetty servers with the specified number of shards and sensible defaults.
-         * Please use the other constructor by passing in the class instance as well.
-         * @param numShards
-         */
-        private JettyServerRule(int numShards)
-        {
-            this.serverName = DEFAULT_TEST_CORENAME;
-            coreNames = new String[]{DEFAULT_TEST_CORENAME};
-            this.numShards = numShards;
-            this.solrcoreProperties = new Properties();
-        }
-
-        /**
-         * Creates the jetty servers with the specified number of shards and sensible defaults.
-         * @param numShards
-         */
-        public JettyServerRule(int numShards, AbstractAlfrescoDistributedTest testClass)
-        {
-            this.serverName = testClass.getClass().getSimpleName();
-            coreNames = new String[]{DEFAULT_TEST_CORENAME};
-            this.numShards = numShards;
-            this.solrcoreProperties = new Properties();
-        }
-        /**
-         * Creates the jetty servers with the specified number of shards and sensible defaults.
-         * @param numShards
-         */
-        public JettyServerRule(int numShards, AbstractAlfrescoDistributedTest testClass, Properties solrcoreProperties)
-        {
-            this.serverName = testClass.getClass().getSimpleName();
-            coreNames = new String[]{DEFAULT_TEST_CORENAME};
-            this.numShards = numShards;
-            this.solrcoreProperties = solrcoreProperties;
-        }
-
-        @Override
-        protected void before() throws Throwable
-        {
-            distribSetUp(serverName);
-            RandVal.uniqueValues = new HashSet(); // reset random values
-            createServers(serverName, coreNames, numShards,solrcoreProperties);
-        }
-
-        @Override
-        protected void after()
-        {
-
-            try
-            {
-                destroyServers();
-                distribTearDown();
-
-                boolean keepTests = Boolean.valueOf(System.getProperty("keep.tests"));
-                if (!keepTests) FileUtils.deleteDirectory(testDir);
-            }
-            catch (Exception e)
-            {
-                log.error("Failed to shutdown test properly ", e);
-            }
-        }
-    }
-
-    /**
-     * Creates a Jetty instance with the default "alfresco" core that uses the production rerank template.
-     * There is only 1 shard.
-     */
-    public class DefaultAlrescoCoreRule extends JettyServerRule
-    {
-        SolrCore defaultCore;
-        SolrClient defaultClient;
-        Properties properties;
-
-        public DefaultAlrescoCoreRule(String serverName, Properties properties) {
-            super(serverName, 0, null, null);
-            this.properties = properties;
-        }
-
-        @Override
-        protected void before() throws Throwable {
-            super.before();
-
-            JettySolrRunner jsr = jettyContainers.get(serverName);
-            CoreContainer coreContainer = jsr.getCoreContainer();
-            AlfrescoCoreAdminHandler coreAdminHandler = (AlfrescoCoreAdminHandler)  coreContainer.getMultiCoreHandler();
-            assertNotNull(coreAdminHandler);
-            String[] extras = null;
-            if ((properties != null) && !properties.isEmpty())
-            {
-                int i = 0;
-                extras = new String[properties.size()*2];
-                for (Map.Entry<Object, Object> prop:properties.entrySet()) {
-                    extras[i++] = "property."+prop.getKey();
-                    extras[i++] = (String) prop.getValue();
-                }
-
-            }
-            defaultCore = createCoreUsingTemplate(coreContainer, coreAdminHandler, "alfresco", "rerank", 1, 1, extras);
-            assertNotNull(defaultCore);
-            String url = buildUrl(jsr.getLocalPort()) + "/" + "alfresco";
-            defaultClient = createNewSolrClient(url);
-            assertNotNull(defaultClient);
-            collectionToStandaloneClient.put("alfresco", defaultClient);
-
-        }
-
-        public SolrCore getDefaultCore() {
-            return defaultCore;
-        }
-
-        public SolrClient getDefaultClient() {
-            return defaultClient;
-        }
     }
 
     public static class BasicAuthFilter implements Filter {
